Remove the dead legacy rule-parsing path (LineParser, SimpleCommonRulesParser & co)
SimpleCommonRulesRewriterFactory - the live production rewriter factory -
has used the querqy.rewrite.rules.* parsing pipeline (RulesParser) since
3.13.0. That left an entire legacy subsystem with no production callers,
only tests: SimpleCommonRulesParser, LineParser's rule-line parsing
entry points, PropertiesBuilder, InputString, ValidationError, and
RulesCollection/TrieMapRulesCollection (production builds a TrieMap and
consumes it via TrieMapLookupQueryVisitorFactory directly, never through
RulesCollectionBuilder#build()).

LineParser's term/input parsing utilities (parseInput, parseTerm,
parseFieldNames, wildcard/boundary validation) were not dead - Input and
InputParserAdapter still use them - but were bundled under the same
class-level @Deprecated. Extracted them into a new InputTermParser,
throwing querqy.rewrite.rules.RuleParseException on invalid input
instead of returning a ValidationError sentinel object, which also let
Input.parseSimpleInput and InputParserAdapter drop their instanceof
plumbing (InputParserAdapter itself stays - it's the real production
adapter RuleParser uses to dispatch simple vs. boolean input, not just
a LineParser bridge).

BOUNDARY, WILDCARD and COMMENT_START moved into a new SpecialChars
class - the syntax definition of "which characters are special" is
logically prior to escaping, so EscapeUtil now consumes these constants
rather than owning them (previously it awkwardly imported them from
LineParser/SimpleCommonRulesParser instead). ESCAPE moved there too for
the same reason, alongside the others it's defined next to.

BOUNDARY_WORD moved from TrieMapRulesCollection into
TrieMapRulesCollectionBuilder, its one remaining live consumer besides
InputSequenceNormalizer. Also removed a stale @Deprecated on
TrieMapRulesCollectionBuilder's lookupPreprocessor field - the comment
claimed it was only kept for the now-removed build() method, but it's
actually used by addSuffixWildcardRule/suffixKeys too.

Checked querqy-solr, querqy-elasticsearch, querqy-opensearch and
querqy-unplugged: none reference any of the removed classes.
