--- querqy-core/src/main/java/querqy/rewriter/commonrules/ValidationError.java
@@ -1,68 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * Copyright 2014 Querqy Contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package querqy.rewriter.commonrules;
-
-/**
- * @author René Kriegler, @renekrie
- *
- */
-@Deprecated
-public class ValidationError {
-
-   final String message;
-
-   public ValidationError(String message) {
-      this.message = message;
-   }
-
-   public String getMessage() {
-      return message;
-   }
-
-   @Override
-   public String toString() {
-      return "ValidationError[message='" + message + "']";
-   }
-
-   @Override
-    public int hashCode() {
-        final int prime = 31;
-        int result = 1;
-        result = prime * result + ((message == null) ? 0 : message.hashCode());
-        return result;
-   }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj)
-            return true;
-        if (obj == null)
-            return false;
-        if (getClass() != obj.getClass())
-            return false;
-        ValidationError other = (ValidationError) obj;
-        if (message == null) {
-            if (other.message != null)
-                return false;
-        } else if (!message.equals(other.message))
-            return false;
-        return true;
-    }
-   
-   
-}
