--- querqy-core/src/main/java/querqy/rewriter/commonrules/model/RulesCollection.java
@@ -1,59 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * Copyright 2014 Querqy Contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package querqy.rewriter.commonrules.model;
-
-import java.util.Set;
-import java.util.stream.Collectors;
-
-import querqy.model.InputSequenceElement;
-import querqy.rewriter.commonrules.select.TopRewritingActionCollector;
-
-@Deprecated
-public interface RulesCollection {
-
-    /** 
-     * Find and return all rewrite actions for an input sequence
-     * @param sequence The input sequence
-     * @param collector The collector of rewriting actions
-     */
-    void collectRewriteActions(PositionSequence<InputSequenceElement> sequence, TopRewritingActionCollector collector);
-    
-    /**
-     * 
-     * Get a collection of all instructions from the rules of this RulesCollection
-     * 
-     * @return The set of instructions
-     */
-    Set<Instruction> getInstructions();
-
-    /**
-     * <p>Get all {@link querqy.model.Term}s that will be generated by this RulesCollection. This only includes terms
-     * that are known from the configuration. The terms returned from this method might be used to optimise caching.</p>
-     *
-     * @return The terms generated by the right-hand sides of the instructions.
-     */
-    default Set<querqy.model.Term> getGenerableTerms() {
-
-        return getInstructions().stream()
-                .flatMap(instruction -> instruction.getGenerableTerms().stream())
-                .collect(Collectors.toSet());
-
-    }
-    
-
-}
