+++ querqy-core/src/main/java/querqy/rewriter/commonrules/model/RulesCollectionBuilder.java
@@ -31,8 +31,6 @@ public interface RulesCollectionBuilder {
 
     void addRule(final Rule rule);
 
-    RulesCollection build();
-
     TrieMap<InstructionsSupplier> getTrieMap();
 
     SuffixWildcardRules<InstructionsSupplier> getSuffixWildcardRules();
